automatically pull nginx version
One of the challenges is knowing which version of nginx is used for each ingress controller version. You need to know that in order to download the correct SigSci nginx module. This Dockerfile will pull out the nginx version and feed it into the SigSci module file download.  As long as the ingress controller has a nginx version that has a corresponding SigSci nginx module, the container should build successfully.  Let me know if you have any issues or questions.